requirements.txt
# LLM
openai>=1.0.0              # OpenAI SDK (兼容DeepSeek)

# 数据库
redis>=5.0.0               # Redis客户端
psycopg2-binary>=2.9.0     # PostgreSQL
sqlalchemy>=2.0.0          # ORM
chromadb>=0.4.0            # 向量数据库

# 工具
python-dotenv>=1.0.0       # 环境变量
tiktoken>=0.5.0            # Token计数
pydantic>=2.0.0            # 数据验证
loguru>=0.7.0              # 日志

# API（可选，后期用）
# fastapi>=0.104.0
# uvicorn>=0.24.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0